requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-docx
pypdf
openai
python-jose
passlib[bcrypt]
